fix(pts): dtype-safe species scatter — masked_block_relabel accepts int8 species (swap-MC ref path)
The s dtype inconsistency (int8 swap-MC ref vs int64 PT2) crashed the T=0.8/0.65 campaign at
masked_block_relabel's scatter (src.dtype != self.dtype). Fixed: scatter casts src to s.dtype;
get_references stores int64; run_cell casts to long. Regression test with int8 species. Both
reviewers flagged this as Minor 'leave' — it was load-bearing for the swap-MC reference path
(untested because tests used PT2/int64 or table_fn=None).

+++ liquid_coupling_flow/ka_pin.py
@@ -76,7 +76,7 @@ def masked_block_relabel(x, s, U, mobile, beta, energy_fn, table_fn, k):
     wblk = W.gather(1, blk); sblk = s.gather(1, blk)
     m = sblk.sum(1).long()
     sblk_new = _cb_sample(wblk, m)
-    s_prop = s.scatter(1, blk, sblk_new)
+    s_prop = s.scatter(1, blk, sblk_new.to(s.dtype))                  # dtype-safe (s may be int8 or int64)
     U_prop = energy_fn(x, s_prop)
     log_ratio = -beta * (U_prop - U) + _cb_logprob(wblk, sblk) - _cb_logprob(wblk, sblk_new)
     acc = torch.log(torch.rand(B, device=x.device)) < log_ratio
